Shut up warning

+++ src/subs_edit_ctrl.cpp
@@ -283,7 +283,7 @@ void SubsTextEditCtrl::SetTextTo(std::string const& text) {
 	Freeze();
 
 	auto insertion_point = GetInsertionPoint();
-	if (insertion_point > line_text.size())
+	if (static_cast<size_t>(insertion_point) > line_text.size())
 		line_text = GetTextRaw().data();
 	auto old_pos = agi::CharacterCount(line_text.begin(), line_text.begin() + insertion_point, 0);
 	line_text.clear();
